Refactoring ListItem classes into their source files. Adding test coverage for EmptyListItem.
