+++ app/src/main/kotlin/com/jameskbride/codemashcompanion/sessions/list/SessionToListItemConverter.kt
@@ -2,6 +2,10 @@ package com.jameskbride.codemashcompanion.sessions.list
 
 import android.view.View
 import com.jameskbride.codemashcompanion.data.model.FullSession
+import com.jameskbride.codemashcompanion.sessions.list.listitems.DateHeaderListItem
+import com.jameskbride.codemashcompanion.sessions.list.listitems.EmptyListItem
+import com.jameskbride.codemashcompanion.sessions.list.listitems.SessionListItem
+import com.jameskbride.codemashcompanion.sessions.list.listitems.TimeHeaderListItem
 import com.xwray.groupie.Group
 
 class SessionToListItemConverter {
